move scheduler into its own file
