+++ src/concurrency/mod.rs
@@ -3,6 +3,7 @@ pub mod cpu_affinity;
 pub mod data_race;
 mod data_race_handler;
 pub mod init_once;
+pub mod scheduler;
 pub mod sync;
 pub mod thread;
 mod vector_clock;
